fix(lucy_config_pipeline): rename python package dir to src and update imports
